Update the boilerplate for the external app
I like to keep everything for this root app in one place. Models are
API-specific too, so it's best to keep models with the APIs.
